+++ packages/observable-client/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - ensure `bestBlocks$` always start with latest known `finalizedBlock` [#491](https://github.com/polkadot-api/polkadot-api/pull/491)
+- Stop-recovery: blocks from previous session getting unpinned
 
 ## 0.2.2 - 2024-05-03